# opt: reduce allocations in JSON value hashing