--- a/pkg/col/coldataext/datum_vec.go
+++ b/pkg/col/coldataext/datum_vec.go
@@ -142,9 +142,12 @@ func (dv *datumVec) Cap() int {
 }
 
 // MarshalAt implements coldata.DatumVec interface.
-func (dv *datumVec) MarshalAt(appendTo []byte, i int) ([]byte, error) {
+func (dv *datumVec) MarshalAt(appendTo []byte, i int) (_ []byte, err error) {
 	dv.maybeSetDNull(i)
-	return valueside.EncodeWithScratch(appendTo, valueside.NoColumnID, dv.data[i], dv.scratch)
+	appendTo, dv.scratch, err = valueside.EncodeWithScratch(
+		appendTo, valueside.NoColumnID, dv.data[i], dv.scratch[:0],
+	)
+	return appendTo, err
 }
 
 // UnmarshalTo implements coldata.DatumVec interface.