--- a/pkg/col/colserde/record_batch_test.go
+++ b/pkg/col/colserde/record_batch_test.go
@@ -227,7 +227,7 @@ func randomDataFromType(rng *rand.Rand, t *types.T, n int, nullProbability float
 		)
 		for i := range data {
 			d := randgen.RandDatum(rng, t, false /* nullOk */)
-			data[i], err = valueside.Encode(data[i], valueside.NoColumnID, d, scratch)
+			data[i], err = valueside.EncodeWithScratch(data[i], valueside.NoColumnID, d, scratch)
 			if err != nil {
 				panic(err)
 			}