--- a/pkg/sql/opt/memo/interner.go
+++ b/pkg/sql/opt/memo/interner.go
@@ -250,13 +250,9 @@ func (c *internCache) Add(item interface{}) {
 // interner. To use, first call the init method, then a series of hash methods.
 // The final value is stored in the hash field.
 type hasher struct {
-	// bytes is a scratch byte array used to serialize certain types of values
-	// during hashing and equality testing.
-	bytes []byte
-
-	// bytes2 is a scratch byte array used to serialize certain types of values
-	// during equality testing.
-	bytes2 []byte
+	// bytes, bytes2, and bytes3 are scratch byte arrays used to serialize
+	// certain types of values during hashing and equality testing.
+	bytes, bytes2, bytes3 []byte
 
 	// hash stores the hash value as it is incrementally computed.
 	hash internHash
@@ -268,6 +264,7 @@ func (h *hasher) Init() {
 	*h = hasher{
 		bytes:  h.bytes,
 		bytes2: h.bytes2,
+		bytes3: h.bytes3,
 		hash:   offset64,
 	}
 }
@@ -361,8 +358,6 @@ func (h *hasher) HashDatum(val tree.Datum) {
 		h.HashUint64(uint64(t.PGEpochDays()))
 	case *tree.DTime:
 		h.HashUint64(uint64(*t))
-	case *tree.DJSON:
-		h.HashString(t.String())
 	case *tree.DTuple:
 		// If labels are present, then hash of tuple's static type is needed to
 		// disambiguate when everything is the same except labels.
@@ -377,7 +372,7 @@ func (h *hasher) HashDatum(val tree.Datum) {
 		h.HashString(t.Locale)
 		h.HashString(t.Contents)
 	default:
-		h.bytes = encodeDatum(h.bytes[:0], val)
+		h.bytes, h.bytes3 = encodeDatum(h.bytes[:0], val, h.bytes3[:0])
 		h.HashBytes(h.bytes)
 	}
 }
@@ -891,9 +886,6 @@ func (h *hasher) IsDatumEqual(l, r tree.Datum) bool {
 	case *tree.DTime:
 		rt := r.(*tree.DTime)
 		return uint64(*lt) == uint64(*rt)
-	case *tree.DJSON:
-		rt := r.(*tree.DJSON)
-		return h.IsStringEqual(lt.String(), rt.String())
 	case *tree.DTuple:
 		rt := r.(*tree.DTuple)
 		// Compare datums and then compare static types if nulls or labels
@@ -915,8 +907,8 @@ func (h *hasher) IsDatumEqual(l, r tree.Datum) bool {
 		}
 		return len(lt.Array) != 0 || h.IsTypeEqual(ltyp, rtyp)
 	default:
-		h.bytes = encodeDatum(h.bytes[:0], l)
-		h.bytes2 = encodeDatum(h.bytes2[:0], r)
+		h.bytes, h.bytes3 = encodeDatum(h.bytes[:0], l, h.bytes3[:0])
+		h.bytes2, h.bytes3 = encodeDatum(h.bytes2[:0], r, h.bytes3[:0])
 		return bytes.Equal(h.bytes, h.bytes2)
 	}
 }
@@ -1374,7 +1366,7 @@ func (h *hasher) IsTransactionModesEqual(l, r tree.TransactionModes) bool {
 // Conversely, if two datums are not equivalent, then their encoded bytes will
 // differ. This will panic if the datum cannot be encoded.
 // Notice: DCollatedString does not encode its collation and won't work here.
-func encodeDatum(b []byte, val tree.Datum) []byte {
+func encodeDatum(b []byte, val tree.Datum, scratch []byte) (res []byte, newScratch []byte) {
 	var err error
 
 	// Fast path: encode the datum using table key encoding. This does not always
@@ -1384,13 +1376,13 @@ func encodeDatum(b []byte, val tree.Datum) []byte {
 	if !colinfo.CanHaveCompositeKeyEncoding(val.ResolvedType()) {
 		b, err = keyside.Encode(b, val, encoding.Ascending)
 		if err == nil {
-			return b
+			return b, newScratch
 		}
 	}
 
-	b, err = valueside.Encode(b, valueside.NoColumnID, val)
+	b, scratch, err = valueside.EncodeWithScratch(b, valueside.NoColumnID, val, scratch)
 	if err != nil {
 		panic(err)
 	}
-	return b
+	return b, scratch
 }