--- a/pkg/sql/opt/memo/interner_test.go
+++ b/pkg/sql/opt/memo/interner_test.go
@@ -1061,7 +1061,7 @@ func BenchmarkEncodeDatum(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		for _, d := range datums {
-			encodeDatum(nil, d)
+			encodeDatum(nil, d, nil /* scratch */)
 		}
 	}
 }