--- a/pkg/sql/rowenc/encoded_datum.go
+++ b/pkg/sql/rowenc/encoded_datum.go
@@ -307,7 +307,7 @@ func (ed *EncDatum) Encode(
 	case catenumpb.DatumEncoding_DESCENDING_KEY:
 		return keyside.Encode(appendTo, ed.Datum, encoding.Descending)
 	case catenumpb.DatumEncoding_VALUE:
-		return valueside.Encode(appendTo, valueside.NoColumnID, ed.Datum, nil /* scratch */)
+		return valueside.Encode(appendTo, valueside.NoColumnID, ed.Datum)
 	default:
 		panic(errors.AssertionFailedf("unknown encoding requested %s", enc))
 	}
@@ -366,7 +366,7 @@ func (ed *EncDatum) Fingerprint(
 		}
 		// We must use value encodings without a column ID even if the EncDatum already
 		// is encoded with the value encoding so that the hashes are indeed unique.
-		fingerprint, err = valueside.Encode(appendTo, valueside.NoColumnID, ed.Datum, nil /* scratch */)
+		fingerprint, err = valueside.Encode(appendTo, valueside.NoColumnID, ed.Datum)
 	} else {
 		// For values that are key encodable, using the ascending key.
 		// Note that using a value encoding will not easily work in case when