--- a/pkg/sql/rowenc/valueside/array.go
+++ b/pkg/sql/rowenc/valueside/array.go
@@ -317,7 +317,7 @@ func encodeArrayElement(b []byte, d tree.Datum) ([]byte, error) {
 		}
 		return encoding.EncodeUntaggedBytesValue(b, encoded), nil
 	case *tree.DTuple:
-		return encodeUntaggedTuple(t, b, encoding.NoColumnID, nil)
+		return encodeUntaggedTuple(t, b, nil)
 	case *tree.DTSQuery:
 		encoded := tsearch.EncodeTSQueryPGBinary(nil, t.TSQuery)
 		return encoding.EncodeUntaggedBytesValue(b, encoded), nil