--- a/pkg/sql/rowenc/valueside/tuple.go
+++ b/pkg/sql/rowenc/valueside/tuple.go
@@ -12,23 +12,25 @@ import (
 )
 
 // encodeTuple produces the value encoding for a tuple.
-func encodeTuple(t *tree.DTuple, appendTo []byte, colID uint32, scratch []byte) ([]byte, error) {
+func encodeTuple(
+	t *tree.DTuple, appendTo []byte, colID uint32, scratch []byte,
+) (_, newScratch []byte, err error) {
 	appendTo = encoding.EncodeValueTag(appendTo, colID, encoding.Tuple)
 	return encodeUntaggedTuple(t, appendTo, scratch)
 }
 
 // encodeUntaggedTuple produces the value encoding for a tuple without a value tag.
-func encodeUntaggedTuple(t *tree.DTuple, appendTo []byte, scratch []byte) ([]byte, error) {
+func encodeUntaggedTuple(
+	t *tree.DTuple, appendTo []byte, scratch []byte,
+) (_, newScratch []byte, err error) {
 	appendTo = encoding.EncodeNonsortingUvarint(appendTo, uint64(len(t.D)))
-
-	var err error
 	for _, dd := range t.D {
-		appendTo, err = EncodeWithScratch(appendTo, NoColumnID, dd, scratch)
+		appendTo, scratch, err = EncodeWithScratch(appendTo, NoColumnID, dd, scratch[:0])
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 	}
-	return appendTo, nil
+	return appendTo, scratch, nil
 }
 
 // decodeTuple decodes a tuple from its value encoding. It is the