… pkg/sql/scrub_test.go