#### opt/bench: add case for INSERT with columns computed from JSON

Release note: None

#### sql/rowenc/valueside: remove unused `colID` param in `encodeUntaggedTuple`

Release note: None

#### sql/rowenc/valueside: introduce `EncodeWithScratch`

The `Encode` function has been split into two functions: `Encode` and
`EncodeWithScratch`. They are the same except that the latter accepts a
scratch buffer that is used to encode certain data types. This was
previously allowed in `Encode`. Splitting up the two functions allows
for easier and safer refactoring in future commits.

Release note: None

#### sql/rowenc/valueside: correctly reuse scratch buffer in `EncodeWithScratch`

`EncodeWithScratch` now returns a new scratch buffer, which may have
grown during encoding. Previously, it was impossible to reuse
allocations that grew the scratch slice. `EncodeWithScratch` call sites
have been refactored to correctly reuse the returned scratch slice.

Release note: None

#### opt/memo/interner: use valueside encoding to hash JSON values

Value-side encoding is now used to hash `tree.DJSON` datums in the
interner, which reduces allocations. A third scratch slice has been
added to the interner that is used for this encoding, which helps reduce
allocations further.

Epic: None

Release note: None