@mgartner 2024-11-11
Before/after benchmark of the `json-insert` and new `json-comp-insert` benchmarks:

```
name                                                       old time/op    new time/op    delta
Phases/json-insert/Simple/Parse                              5.13µs ± 2%    5.11µs ± 2%     ~     (p=0.841 n=5+5)
Phases/json-insert/Simple/OptBuildNoNorm                     15.4µs ± 7%    14.8µs ± 1%     ~     (p=0.206 n=5+5)
Phases/json-insert/Simple/OptBuildNorm                       15.1µs ± 1%    14.8µs ± 1%   -2.28%  (p=0.008 n=5+5)
Phases/json-insert/Simple/Explore                            17.4µs ± 2%    16.5µs ± 1%   -4.89%  (p=0.008 n=5+5)
Phases/json-insert/Simple/ExecBuild                          18.1µs ± 2%    17.4µs ± 0%   -3.83%  (p=0.008 n=5+5)
Phases/json-insert/Prepared/ExecBuild                        1.06µs ± 3%    1.02µs ± 1%     ~     (p=0.056 n=5+5)
Phases/json-comp-insert/Simple/Parse                         5.83µs ± 2%    5.84µs ± 1%     ~     (p=0.690 n=5+5)
Phases/json-comp-insert/Simple/OptBuildNoNorm                58.7µs ± 4%    57.0µs ± 4%     ~     (p=0.056 n=5+5)
Phases/json-comp-insert/Simple/OptBuildNorm                  69.7µs ± 3%    66.4µs ± 6%   -4.79%  (p=0.032 n=5+5)
Phases/json-comp-insert/Simple/Explore                       70.2µs ± 3%    67.5µs ± 3%   -3.85%  (p=0.032 n=5+5)
Phases/json-comp-insert/Simple/ExecBuild                     70.3µs ± 1%    68.2µs ± 2%   -3.05%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNoNorm    9.92µs ± 0%    9.63µs ± 1%   -2.90%  (p=0.016 n=4+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNorm      18.4µs ± 2%    15.4µs ± 1%  -16.53%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/Explore                     20.5µs ± 2%    17.3µs ± 1%  -15.46%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/ExecBuild                   21.3µs ± 3%    18.2µs ± 1%  -14.38%  (p=0.008 n=5+5)

name                                                       old alloc/op   new alloc/op   delta
Phases/json-insert/Simple/Parse                              5.42kB ± 0%    5.42kB ± 0%     ~     (all equal)
Phases/json-insert/Simple/OptBuildNoNorm                     15.1kB ± 0%    15.2kB ± 0%   +0.37%  (p=0.008 n=5+5)
Phases/json-insert/Simple/OptBuildNorm                       15.1kB ± 0%    15.2kB ± 0%   +0.37%  (p=0.008 n=5+5)
Phases/json-insert/Simple/Explore                            16.5kB ± 0%    16.6kB ± 0%   +0.34%  (p=0.008 n=5+5)
Phases/json-insert/Simple/ExecBuild                          17.4kB ± 0%    17.4kB ± 0%   +0.32%  (p=0.008 n=5+5)
Phases/json-insert/Prepared/ExecBuild                          896B ± 0%      896B ± 0%     ~     (all equal)
Phases/json-comp-insert/Simple/Parse                         5.52kB ± 0%    5.52kB ± 0%     ~     (all equal)
Phases/json-comp-insert/Simple/OptBuildNoNorm                41.9kB ± 0%    41.6kB ± 0%   -0.82%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/OptBuildNorm                  48.1kB ± 0%    45.6kB ± 0%   -5.15%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/Explore                       49.5kB ± 0%    47.0kB ± 0%   -4.99%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/ExecBuild                     50.4kB ± 0%    47.9kB ± 0%   -4.90%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNoNorm    6.91kB ± 0%    6.56kB ± 0%   -4.97%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNorm      10.8kB ± 0%     8.3kB ± 0%  -22.85%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/Explore                     12.2kB ± 0%     9.8kB ± 0%  -20.22%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/ExecBuild                   13.2kB ± 0%    10.7kB ± 0%  -18.79%  (p=0.008 n=5+5)

name                                                       old allocs/op  new allocs/op  delta
Phases/json-insert/Simple/Parse                                26.0 ± 0%      26.0 ± 0%     ~     (all equal)
Phases/json-insert/Simple/OptBuildNoNorm                       94.0 ± 0%      96.0 ± 0%   +2.13%  (p=0.008 n=5+5)
Phases/json-insert/Simple/OptBuildNorm                         94.0 ± 0%      96.0 ± 0%   +2.13%  (p=0.008 n=5+5)
Phases/json-insert/Simple/Explore                              98.0 ± 0%     100.0 ± 0%   +2.04%  (p=0.008 n=5+5)
Phases/json-insert/Simple/ExecBuild                             105 ± 0%       107 ± 0%   +1.90%  (p=0.008 n=5+5)
Phases/json-insert/Prepared/ExecBuild                          9.00 ± 0%      9.00 ± 0%     ~     (all equal)
Phases/json-comp-insert/Simple/Parse                           34.0 ± 0%      34.0 ± 0%     ~     (all equal)
Phases/json-comp-insert/Simple/OptBuildNoNorm                   260 ± 0%       252 ± 0%   -3.08%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/OptBuildNorm                     353 ± 0%       293 ± 0%  -17.00%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/Explore                          357 ± 0%       297 ± 0%  -16.81%  (p=0.008 n=5+5)
Phases/json-comp-insert/Simple/ExecBuild                        364 ± 0%       304 ± 0%  -16.48%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNoNorm      47.0 ± 0%      39.0 ± 0%  -17.02%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/AssignPlaceholdersNorm         138 ± 0%        78 ± 0%  -43.48%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/Explore                        142 ± 0%        82 ± 0%  -42.25%  (p=0.008 n=5+5)
Phases/json-comp-insert/Prepared/ExecBuild                      149 ± 0%        89 ± 0%  -40.27%  (p=0.008 n=5+5)
```